feat(store): card 1 — raw per-stream export + StreamLister (#145)
Export is a raw pass-through, not a transform. `EventExporter::export_stream`
forwards to `read_stream`, so the associated `ExportStream` is the adapter's
own concrete `S::Stream` — static dispatch, no boxing, no per-event rewrite.
A blanket `impl<S: RawEventStore> EventExporter for S` makes every store an
exporter for free. `StreamLister::list_streams` is the one new capability;
all-streams export is `list_streams` ∘ `export_stream`.

The earlier `PortableEvent` (per-event stream_id, dropped global_seq) is
removed: the caller supplies the id (per-stream, like a read), and import
re-appends through the write path which restamps a fresh global_seq anyway,
so stripping it at export is redundant. The stream id is recorded once per
stream section in the backup box (card 3); import ignores global_seq.

`from` is inclusive (matches read_stream; `Version` is NonZeroU64 so an
exclusive `from` could never include v1). Doc corrected accordingly.

PortableEvent's strong test suite is preserved by migrating it onto
PersistedEnvelope (its validator was a literal copy): zero-copy ptr-equality,
the mutation-killing `end - start` cap pin, never-panics fuzz, full-accessor
proptest, thread-boundary Send. stream_id tests adapted to global_seq; the
circular differential oracle dropped.

import.rs's trait now takes `&[PersistedEnvelope]` (provisional) — card 2
reshapes routing around per-stream sections.

Tests: 16 new export/list_streams across the 4 cross-cutting categories
(sequence, lifecycle, defensive boundary, linearizability) + 19 migrated
envelope tests. Strict clippy clean; fjall + default build unaffected.
