+++ crates/nexus-store/src/envelope.rs
@@ -823,3 +823,653 @@ mod tests {
         assert!(env.metadata_value().is_none());
     }
 }
+
+// ============================================================================
+// Exhaustive PersistedEnvelope suite — migrated from the former PortableEvent
+// tests (issue #145). PortableEvent's validator was a literal copy of
+// PersistedEnvelope's, so its range/UTF-8/cap/zero-copy/fuzz tests apply here
+// unchanged. The stream_id-specific tests were dropped (PersistedEnvelope has
+// global_seq instead); the differential oracle was dropped (it compared the
+// two now-unified validators). These cover the 4 cross-cutting categories:
+// sequence/protocol, lifecycle, defensive boundary, and Send/Sync.
+// ============================================================================
+#[cfg(test)]
+#[allow(clippy::expect_used, reason = "test code")]
+mod persisted_exhaustive_tests {
+    use super::{EnvelopeError, PersistedEnvelope};
+    use crate::store::GlobalSeq;
+    use crate::value::{MAX_EVENT_TYPE_LEN, Metadata, SchemaVersion};
+    use bytes::Bytes;
+    use nexus::Version;
+    use proptest::prelude::*;
+    use static_assertions::assert_impl_all;
+    use std::ops::Range;
+
+    // ── helpers ─────────────────────────────────────────────────────────────
+
+    fn v(n: u64) -> Version {
+        Version::new(n).expect("test version must be nonzero")
+    }
+
+    fn sv(n: u32) -> SchemaVersion {
+        SchemaVersion::from_u32(n).expect("test schema_version must be nonzero")
+    }
+
+    fn gs(n: u64) -> GlobalSeq {
+        GlobalSeq::new(n).expect("test global_seq must be nonzero")
+    }
+
+    /// Assemble a contiguous `[event_type][payload][metadata?]` buffer and the
+    /// matching ranges — the exact shape a decoder hands `try_new`.
+    fn assemble(
+        event_type: &[u8],
+        payload: &[u8],
+        metadata: Option<&[u8]>,
+    ) -> (Bytes, Range<u32>, Range<u32>, Option<Range<u32>>) {
+        let mut buf = Vec::new();
+        buf.extend_from_slice(event_type);
+        buf.extend_from_slice(payload);
+        if let Some(m) = metadata {
+            buf.extend_from_slice(m);
+        }
+        let et_end = u32::try_from(event_type.len()).expect("event_type len fits u32");
+        let pl_end = et_end + u32::try_from(payload.len()).expect("payload len fits u32");
+        let meta_range = metadata.map(|m| {
+            let end = pl_end + u32::try_from(m.len()).expect("metadata len fits u32");
+            pl_end..end
+        });
+        (Bytes::from(buf), 0..et_end, et_end..pl_end, meta_range)
+    }
+
+    /// Build a valid `PersistedEnvelope` from byte components.
+    #[allow(
+        clippy::too_many_arguments,
+        reason = "test helper mirrors the 7-field PersistedEnvelope::try_new shape"
+    )]
+    fn build(
+        version: Version,
+        global_seq: GlobalSeq,
+        schema: SchemaVersion,
+        event_type: &[u8],
+        payload: &[u8],
+        metadata: Option<&[u8]>,
+    ) -> PersistedEnvelope {
+        let (value, et, pl, meta) = assemble(event_type, payload, metadata);
+        PersistedEnvelope::try_new(version, global_seq, value, schema, et, pl, meta)
+            .expect("components form a valid PersistedEnvelope")
+    }
+
+    // ── Category 4: linearizability / Send + Sync ───────────────────────────
+    // PersistedEnvelope flows through `futures::Stream` items across `.await`
+    // points and threads; Send + Sync is a structural contract.
+
+    assert_impl_all!(PersistedEnvelope: Send, Sync, Clone, std::fmt::Debug);
+
+    #[test]
+    fn persisted_envelope_clones_across_thread_boundary() {
+        let event = build(
+            v(3),
+            gs(99),
+            sv(2),
+            b"MoneyDeposited",
+            b"payload-bytes",
+            Some(b"meta"),
+        );
+        let moved = event.clone();
+        // `std::thread::scope` (not the banned free `std::thread::spawn`) moves
+        // the clone to another thread and joins it — proving `Send` in practice.
+        let (gseq, ver, schema, etype, payload, meta) = std::thread::scope(|scope| {
+            scope
+                .spawn(move || {
+                    (
+                        moved.global_seq(),
+                        moved.version(),
+                        moved.schema_version(),
+                        moved.event_type().to_owned(),
+                        moved.payload().to_vec(),
+                        moved.metadata().map(<[u8]>::to_vec),
+                    )
+                })
+                .join()
+                .expect("worker thread must not panic")
+        });
+
+        assert_eq!(gseq, gs(99));
+        assert_eq!(ver, v(3));
+        assert_eq!(schema, 2);
+        assert_eq!(etype, "MoneyDeposited");
+        assert_eq!(payload, b"payload-bytes");
+        assert_eq!(meta, Some(b"meta".to_vec()));
+        // Original still usable after the clone left for another thread.
+        assert_eq!(event.event_type(), "MoneyDeposited");
+    }
+
+    // ── Category 1: sequence / protocol ─────────────────────────────────────
+    // Multi-step interaction on ONE object: repeated and interleaved accessor
+    // calls must be deterministic and mutually consistent.
+
+    #[test]
+    fn repeated_accessor_calls_are_consistent() {
+        let event = build(v(9), gs(5), sv(4), b"TYPE", b"payload", Some(b"meta"));
+
+        // Same accessor twice → identical.
+        assert_eq!(event.event_type(), event.event_type());
+        assert_eq!(event.payload(), event.payload());
+        assert_eq!(event.metadata(), event.metadata());
+
+        // Interleave borrowed and owned views → still consistent.
+        assert_eq!(
+            event.event_type().as_bytes(),
+            event.event_type_bytes().as_ref()
+        );
+        assert_eq!(event.payload(), event.payload_bytes().as_ref());
+        assert_eq!(event.metadata(), event.metadata_bytes().as_deref());
+
+        // Borrowed vs validated-newtype views agree.
+        assert_eq!(event.event_type(), event.event_type_value().as_str());
+        assert_eq!(event.payload(), event.payload_value().as_slice());
+        assert_eq!(
+            event.metadata(),
+            event.metadata_value().as_ref().map(Metadata::as_slice),
+        );
+
+        // Scalars are stable across repeated reads.
+        assert_eq!(event.version(), v(9));
+        assert_eq!(event.version(), event.version());
+        assert_eq!(event.global_seq(), gs(5));
+        assert_eq!(event.global_seq(), event.global_seq());
+        assert_eq!(event.schema_version(), 4);
+        assert_eq!(event.schema_version_value(), sv(4));
+    }
+
+    // ── Category 2: lifecycle (build / clone / access-after-clone) ───────────
+
+    #[test]
+    fn clone_is_field_for_field_equal_view() {
+        let original = build(
+            v(42),
+            gs(7),
+            sv(7),
+            b"OrderPlaced",
+            b"the-payload",
+            Some(b"the-meta"),
+        );
+        let cloned = original.clone();
+
+        assert_eq!(cloned.version(), original.version());
+        assert_eq!(cloned.global_seq(), original.global_seq());
+        assert_eq!(cloned.schema_version(), original.schema_version());
+        assert_eq!(
+            cloned.schema_version_value(),
+            original.schema_version_value()
+        );
+        assert_eq!(cloned.event_type(), original.event_type());
+        assert_eq!(cloned.payload(), original.payload());
+        assert_eq!(cloned.metadata(), original.metadata());
+        assert_eq!(
+            cloned.metadata_value().map(|m| m.as_slice().to_vec()),
+            original.metadata_value().map(|m| m.as_slice().to_vec()),
+        );
+    }
+
+    #[test]
+    fn clone_shares_the_same_backing_buffer() {
+        // A clone is one Arc refcount inc — owned views from both clones must
+        // point at the SAME allocation (zero-copy), not a deep copy.
+        let original = build(v(1), gs(1), sv(1), b"TYPE", b"payload", None);
+        let cloned = original.clone();
+
+        let from_original = original.payload_bytes();
+        let from_clone = cloned.payload_bytes();
+        assert!(
+            std::ptr::eq(from_original.as_ptr(), from_clone.as_ptr()),
+            "clone must share the parent buffer, not deep-copy it",
+        );
+        assert_eq!(from_original.as_ref(), from_clone.as_ref());
+    }
+
+    // ── Category 3: defensive boundary (reject upstream-guarantee violations) ─
+
+    #[test]
+    fn rejects_inverted_range_start_after_end() {
+        let value = Bytes::from_static(b"TYPEpayload");
+        // Built from variables so the inverted range is not a compile-time
+        // literal (which `reversed_empty_ranges` would reject before runtime).
+        let (bad_start, bad_end) = (7u32, 2u32);
+        let err = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            0..4,
+            bad_start..bad_end,
+            None,
+        )
+        .expect_err("start > end must be rejected");
+        assert!(matches!(
+            err,
+            EnvelopeError::RangeOutOfBounds {
+                start: 7,
+                end: 2,
+                ..
+            }
+        ));
+    }
+
+    #[test]
+    fn event_type_cap_uses_range_length_not_endpoint_sum() {
+        // Mutation-testing pin (cargo-mutants surfaced this gap): the cap check
+        // is on `event_type_range.end - event_type_range.start`, NOT `+`. Every
+        // other test starts the event_type range at 0, where `end - 0 == end +
+        // 0`, so none can tell the operators apart. Here start > 0 and the
+        // length (end - start = 30_000) is WITHIN the cap, while the endpoint
+        // sum (end + start = 110_000) EXCEEDS it — so a `- → +` mutant would
+        // wrongly reject this valid event, and the `.expect` below catches it.
+        let start: u32 = 40_000;
+        let end: u32 = 70_000;
+        let buf_len = usize::try_from(end).expect("70_000 fits usize");
+        let value = Bytes::from(vec![b'A'; buf_len]);
+        let event = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            start..end,
+            0..0,
+            None,
+        )
+        .expect("event_type length 30_000 is within MAX_EVENT_TYPE_LEN (65_535)");
+        assert_eq!(event.event_type().len(), 30_000);
+    }
+
+    // Each of the three ranges is bounds-checked independently. The payload
+    // path is covered in the outer module; these pin the event_type and
+    // metadata paths so a regression deleting either check is caught.
+
+    #[test]
+    fn rejects_event_type_range_past_buffer_end() {
+        let value = Bytes::from_static(b"short");
+        let err = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            // event_type is the FIRST range checked; end 100 > len 5.
+            0..100,
+            0..0,
+            None,
+        )
+        .expect_err("event_type range past buffer must be rejected");
+        assert!(matches!(
+            err,
+            EnvelopeError::RangeOutOfBounds {
+                start: 0,
+                end: 100,
+                len: 5
+            }
+        ));
+    }
+
+    #[test]
+    fn rejects_metadata_range_past_buffer_end() {
+        let value = Bytes::from_static(b"TYPEpayload");
+        let err = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            // event_type + payload in-bounds so the metadata check is reached.
+            0..4,
+            4..11,
+            Some(11..100),
+        )
+        .expect_err("metadata range past buffer must be rejected");
+        assert!(matches!(
+            err,
+            EnvelopeError::RangeOutOfBounds {
+                start: 11,
+                end: 100,
+                len: 11
+            }
+        ));
+    }
+
+    #[test]
+    fn rejects_inverted_metadata_range() {
+        let value = Bytes::from_static(b"TYPEpayload");
+        let (bad_start, bad_end) = (9u32, 5u32);
+        let err = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            0..4,
+            4..11,
+            Some(bad_start..bad_end),
+        )
+        .expect_err("metadata start > end must be rejected");
+        assert!(matches!(
+            err,
+            EnvelopeError::RangeOutOfBounds {
+                start: 9,
+                end: 5,
+                ..
+            }
+        ));
+    }
+
+    // ── absence-of-invariant pins: independent (possibly overlapping) ranges ─
+
+    #[test]
+    fn ranges_are_independent_and_may_overlap() {
+        // try_new enforces NO disjointness/contiguity. All three windows can
+        // point at the SAME 4 bytes. A future change adding a disjointness
+        // invariant is a conscious break, caught here.
+        let value = Bytes::from_static(b"TYPE");
+        let event = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            0..4,
+            0..4,
+            Some(0..4),
+        )
+        .expect("overlapping ranges are structurally permitted");
+        assert_eq!(event.event_type(), "TYPE");
+        assert_eq!(event.payload(), b"TYPE");
+        assert_eq!(event.metadata(), Some(b"TYPE".as_slice()));
+    }
+
+    #[test]
+    fn empty_payload_range_is_accepted_unlike_empty_metadata() {
+        // Deliberate asymmetry mirrored from the value newtypes: an empty
+        // payload is a legal event shape (marker event), an empty metadata
+        // Some is not (use None).
+        let value = Bytes::from_static(b"TYPE");
+        let event = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            0..4,
+            4..4,
+            None,
+        )
+        .expect("empty payload range is accepted");
+        assert_eq!(event.payload(), b"");
+        assert_eq!(event.payload_value().as_slice(), b"");
+    }
+
+    // ── metadata None vs Some across every metadata accessor ────────────────
+
+    #[test]
+    fn metadata_absent_is_none_everywhere() {
+        let event = build(v(1), gs(1), sv(1), b"TYPE", b"payload", None);
+        assert!(event.metadata().is_none());
+        assert!(event.metadata_bytes().is_none());
+        assert!(event.metadata_value().is_none());
+    }
+
+    #[test]
+    fn metadata_present_threads_through_every_accessor() {
+        let event = build(v(1), gs(1), sv(1), b"TYPE", b"payload", Some(b"META"));
+        assert_eq!(event.metadata(), Some(b"META".as_slice()));
+        assert_eq!(event.metadata_bytes().expect("some").as_ref(), b"META");
+        assert_eq!(event.metadata_value().expect("some").as_slice(), b"META");
+    }
+
+    // ── empty event_type accepted (no minimum-length invariant) ─────────────
+
+    #[test]
+    fn empty_event_type_is_accepted() {
+        let event = build(v(1), gs(1), sv(1), b"", b"payload", None);
+        assert_eq!(event.event_type(), "");
+        assert_eq!(event.event_type_value().as_str(), "");
+        assert_eq!(event.payload(), b"payload");
+    }
+
+    // ── boundary scalars carried verbatim (version, global_seq, schema) ──────
+
+    #[test]
+    fn boundary_version_global_seq_and_schema_version_carried_verbatim() {
+        let event = build(
+            v(u64::MAX),
+            gs(u64::MAX),
+            sv(u32::MAX),
+            b"TYPE",
+            b"payload",
+            None,
+        );
+        assert_eq!(event.version().as_u64(), u64::MAX);
+        assert_eq!(event.global_seq().as_u64(), u64::MAX);
+        assert_eq!(event.schema_version(), u32::MAX);
+        assert_eq!(event.schema_version_value().get(), u32::MAX);
+    }
+
+    // ── zero-copy aliasing: owned views share the one Arc buffer ────────────
+
+    #[test]
+    fn owned_views_alias_the_single_backing_buffer() {
+        let (value, et, pl, meta) = assemble(b"TYPE", b"payload", Some(b"meta"));
+        let base = value.clone(); // shares the same allocation as `value`
+        let event = PersistedEnvelope::try_new(
+            Version::INITIAL,
+            gs(1),
+            value,
+            SchemaVersion::INITIAL,
+            et,
+            pl,
+            meta,
+        )
+        .expect("valid");
+
+        let et_bytes = event.event_type_bytes();
+        let pl_bytes = event.payload_bytes();
+        let meta_bytes = event.metadata_bytes().expect("metadata present");
+
+        // Each owned view points into `base` at exactly its range offset →
+        // proves no copy AND that all three share the one buffer.
+        assert!(
+            std::ptr::eq(et_bytes.as_ptr(), base.as_ptr().wrapping_add(0)),
+            "event_type_bytes must alias buffer offset 0",
+        );
+        assert!(
+            std::ptr::eq(pl_bytes.as_ptr(), base.as_ptr().wrapping_add(4)),
+            "payload_bytes must alias buffer offset 4",
+        );
+        assert!(
+            std::ptr::eq(meta_bytes.as_ptr(), base.as_ptr().wrapping_add(11)),
+            "metadata_bytes must alias buffer offset 11",
+        );
+        // The value-newtype path shares the same buffer too.
+        assert!(std::ptr::eq(
+            event.event_type_value().as_bytes().as_ptr(),
+            base.as_ptr().wrapping_add(0),
+        ));
+        assert!(std::ptr::eq(
+            event.payload_value().as_slice().as_ptr(),
+            base.as_ptr().wrapping_add(4),
+        ));
+    }
+
+    // ── Property: random valid components round-trip through every accessor ──
+
+    fn boundary_u64() -> impl Strategy<Value = u64> {
+        prop_oneof![
+            Just(1u64),
+            Just(2u64),
+            Just(u64::MAX - 1),
+            Just(u64::MAX),
+            1u64..=u64::MAX,
+        ]
+    }
+
+    fn boundary_u32_nonzero() -> impl Strategy<Value = u32> {
+        prop_oneof![
+            Just(1u32),
+            Just(2u32),
+            Just(u32::MAX - 1),
+            Just(u32::MAX),
+            1u32..=u32::MAX,
+        ]
+    }
+
+    fn event_type_bytes_strategy() -> impl Strategy<Value = Vec<u8>> {
+        // ASCII 'A'..='Z' guarantees valid UTF-8; lengths include the empty,
+        // unit, interior, and exact-cap boundaries.
+        prop_oneof![
+            Just(0usize),
+            Just(1usize),
+            Just(255usize),
+            Just(MAX_EVENT_TYPE_LEN),
+        ]
+        .prop_flat_map(|n| proptest::collection::vec(b'A'..=b'Z', n))
+    }
+
+    fn payload_bytes_strategy() -> impl Strategy<Value = Vec<u8>> {
+        prop_oneof![Just(0usize), Just(1usize), Just(1024usize)]
+            .prop_flat_map(|n| proptest::collection::vec(any::<u8>(), n))
+    }
+
+    fn metadata_bytes_strategy() -> impl Strategy<Value = Option<Vec<u8>>> {
+        prop_oneof![
+            Just(None),
+            Just(Some(vec![0u8])),
+            (1usize..=2048)
+                .prop_flat_map(|n| proptest::collection::vec(any::<u8>(), n))
+                .prop_map(Some),
+        ]
+    }
+
+    proptest! {
+        #[test]
+        fn persisted_envelope_roundtrips_every_accessor(
+            event_type in event_type_bytes_strategy(),
+            payload in payload_bytes_strategy(),
+            metadata in metadata_bytes_strategy(),
+            version_raw in boundary_u64(),
+            global_seq_raw in boundary_u64(),
+            schema_raw in boundary_u32_nonzero(),
+        ) {
+            let version = v(version_raw);
+            let global_seq = gs(global_seq_raw);
+            let schema = sv(schema_raw);
+            let event = build(
+                version,
+                global_seq,
+                schema,
+                &event_type,
+                &payload,
+                metadata.as_deref(),
+            );
+
+            // Scalars verbatim.
+            prop_assert_eq!(event.version(), version);
+            prop_assert_eq!(event.version().as_u64(), version_raw);
+            prop_assert_eq!(event.global_seq(), global_seq);
+            prop_assert_eq!(event.global_seq().as_u64(), global_seq_raw);
+            prop_assert_eq!(event.schema_version(), schema_raw);
+            prop_assert_eq!(event.schema_version_value(), schema);
+
+            // event_type across all three views. Owned views bound to locals
+            // first so the backing Bytes/newtype outlives the borrow.
+            let et_bytes = event.event_type_bytes();
+            let et_value = event.event_type_value();
+            prop_assert_eq!(event.event_type().as_bytes(), event_type.as_slice());
+            prop_assert_eq!(et_bytes.as_ref(), event_type.as_slice());
+            prop_assert_eq!(et_value.as_bytes(), event_type.as_slice());
+
+            // payload across all three views.
+            let pl_bytes = event.payload_bytes();
+            let pl_value = event.payload_value();
+            prop_assert_eq!(event.payload(), payload.as_slice());
+            prop_assert_eq!(pl_bytes.as_ref(), payload.as_slice());
+            prop_assert_eq!(pl_value.as_slice(), payload.as_slice());
+
+            // metadata: presence and bytes consistent across all three views.
+            if let Some(ref m) = metadata {
+                prop_assert_eq!(event.metadata(), Some(m.as_slice()));
+                let meta_bytes = event.metadata_bytes().expect("some");
+                prop_assert_eq!(meta_bytes.as_ref(), m.as_slice());
+                let meta_value = event.metadata_value().expect("some");
+                prop_assert_eq!(meta_value.as_slice(), m.as_slice());
+            } else {
+                prop_assert!(event.metadata().is_none());
+                prop_assert!(event.metadata_bytes().is_none());
+                prop_assert!(event.metadata_value().is_none());
+            }
+        }
+    }
+
+    // ── Adversarial robustness: try_new is fed UNTRUSTED decoder output ──────
+    //
+    // On the read path a corrupt/adversarial row reaches try_new. It must be
+    // panic-free for ANY (value, ranges) and either return Ok or a KNOWN
+    // EnvelopeError — never a panic, and never an Ok whose unsafe accessors are
+    // unsound. The exhaustive Err match has no catch-all, so a new variant is a
+    // compile error here — pinning the rejection surface. Run under Miri to
+    // prove the unsafe accessors' preconditions hold for fuzzed input.
+
+    fn arbitrary_range() -> impl Strategy<Value = Range<u32>> {
+        (0u32..=260, 0u32..=260).prop_map(|(start, end)| start..end)
+    }
+
+    fn arbitrary_optional_range() -> impl Strategy<Value = Option<Range<u32>>> {
+        prop_oneof![Just(None), arbitrary_range().prop_map(Some)]
+    }
+
+    proptest! {
+        #[test]
+        fn try_new_never_panics_and_accepted_events_have_sound_accessors(
+            value_bytes in proptest::collection::vec(any::<u8>(), 0..256),
+            event_type_range in arbitrary_range(),
+            payload_range in arbitrary_range(),
+            metadata_range in arbitrary_optional_range(),
+            version_raw in boundary_u64(),
+            global_seq_raw in boundary_u64(),
+            schema_raw in boundary_u32_nonzero(),
+        ) {
+            let result = PersistedEnvelope::try_new(
+                v(version_raw),
+                gs(global_seq_raw),
+                Bytes::from(value_bytes),
+                sv(schema_raw),
+                event_type_range,
+                payload_range,
+                metadata_range,
+            );
+
+            match result {
+                Ok(event) => {
+                    // Every accessor must be callable without panic/UB. The
+                    // event_type() / *_value() paths run `unsafe`
+                    // from_utf8_unchecked / from_validated_bytes — under Miri
+                    // this proves their preconditions held for fuzzed input.
+                    let et = event.event_type();
+                    prop_assert!(std::str::from_utf8(et.as_bytes()).is_ok());
+                    let _ = event.payload();
+                    let _ = event.metadata();
+                    let _ = event.event_type_value();
+                    let _ = event.payload_value();
+                    let _ = event.metadata_value();
+                    let _ = event.event_type_bytes();
+                    let _ = event.payload_bytes();
+                    let _ = event.metadata_bytes();
+                }
+                Err(err) => {
+                    // Exhaustive — no `_` arm. A new EnvelopeError variant must
+                    // be classified here, not silently absorbed.
+                    match err {
+                        EnvelopeError::RangeOutOfBounds { .. }
+                        | EnvelopeError::InvalidUtf8 { .. }
+                        | EnvelopeError::EventTypeRangeTooLong { .. }
+                        | EnvelopeError::MetadataRangeTooLong { .. }
+                        | EnvelopeError::MetadataRangeEmpty
+                        | EnvelopeError::Value(_) => {}
+                    }
+                }
+            }
+        }
+    }
+}
